Create parent class for MipUpload and child classes for dna and rna
